Rename to "mediawiki/phan-security-plugin"
The mediawiki vendor is to match our other CI tools that use that namespace,
and also because the main intention is that this tool will be used for
MediaWiki.

Since this is a phan plugin, it should mention phan in the package name. I
dropped "check" since that was somewhat generic and implied.

Change-Id: If5e46d906aa4cb399d76125e25b56ceff0d2b02a