Add support for docblock taint annotations
Instead of modifying the plugin to tell it about custom
taint information, the docblock comment for the method can be changed.
This should allow phan to better integrate with running against
different branches of the code where taint info can change between
branches.

For example, a method that escapes html can now have in the docblock:

  * @param-taint $html escapes_html

Which would tell phan-taint-check that any html put into that
parameter comes out the other end of the method safe for outputting
into the document.

Thanks to Legoktm for the idea.

Bug: T187377
Change-Id: Ib34499ecb165122a2628c21f4770168544660b93