Improve tracking of outputting class members
This ensures that if a method does something like
	return $this->foo;
and elsewhere we have:
	$this->setFoo( $somethingEvil );
	echo $this->getFoo();
That we properly emit a warning.

This may have some false positives, as it will conflate
all instances of a class with each other. The assumption being
that its bad practise to use a class sometimes for already escaped
variables and other times for dangerous variables.

This is accomplished two ways:
* Ensuring that if a method's taint depends on a class member, that
  the cached value for the method's taint is cleared any time the
  taint of a class member is updated (e.g. If a setter is called with
  a tainted value)
* If a method directly returns a class member (i.e. is a getter), and the
  getter is directly output, we mark the class member as something
  that is outputted (This only comes into play if we encounter the
  output code priort to the set tainted value code, and is less robust
  than the other code path)

Change-Id: I630697b092775f7d6af3a3d27b7b2d231ed4406f