Fix bug in how taint propagation works
If inside a branch you called a function with a semi-tainted
variable as its argument, and the formal parameter name used by the
function was the same as the variable name of the argument, then
the variable becomes fully tainted.

This was caused by two bugs:
* When setting taintedness of a variable, setTaintedness was looking
  in the current context's scope, not the variable's scope, for parent variables.
* markAllDependentVarsYes() marked things as YES_TAINT instead of the specific
  taint that would be appropriate.

Additionally, this changes the definition of ALL_TAINT to be defined
in terms of other constants instead of a literal number, because I was
confusing myself.

This fixes a false positive in Wikibase EntityUsageTable.php. It also fixes
the wfDebug() false positives in MW core.

Change-Id: I0f303bde592f53b22ed42160ae8a956c4a7b1706