Prevent an EXEC variable from tainting itself
If a variable is marked EXEC (e.g. it is echoed out later), and
it is marked tainted, previously when assigning it to another
variable, and the method is processed more than once, it would
result in a false positive, because it would appear as if the
variable is tainting itself the second time around.

Change-Id: I6fae534b534e15ec485e73be3895d89c5e95c0e0
(cherry picked from commit 1946422f0ab005293efb640a684c99a4a3b94219)