When suppressing a warning, also suppress side effects
When something unsafe is done with a variable, it is marked as
unsafe to assign to, in case the evil assignment hasn't happened
yet. This means that when suppressing a warning, the warning might
not be totally suppressed as the variable is still marked as unsafe
to assign to. Now the variable is no longer marked as such.

This also fixes a typo that prevents double escaping status
from being propagated in some circumstances. In practise this
seems to increase the number of double escape warnings by
several hundred in MediaWiki core.

Change-Id: Id18f234c4643a44a8f920591c993ab673efa8305