Avoid false positives with polymorphic methods taking string|Message
Disable double escape checking for methods taking a string|Message
as the tool gets confused due to Message::__toString().

Ideally we would be able to treat methods with different type
signatures differently, and have smart treatment of conditionals
that apply only to one type, but that is a lot harder and this
solves the immediate problem.

Bug: T269895
Change-Id: I94cf0cc9be01385fb80dc41ee63d979ca46eb4db