Upgrade phan to 0.8.6
The first small step to 1.2.x. This notably allows seccheck to be run on
PHP 7.1.
Most of the changes are pretty trivial/clean, aside from the one for
getPropertyByNameInContext. I really couldn't find a better way to
change it. However, I saw that newer versions of phan have different
methods to get class properties, so those could be a solution as we'll
get to a more updated version.

Note: I tested it locally on core and CheckUser, and it seemed to work
on both (although with many issues found). Then I tried to test it on AbuseFilter (where seccheck currently pass) but went OOM :/. I also ran runtest.sh and it seems good, aside from a few Windows incompatibilities.

Bug: T216974
Depends-On: I79f7120f015b7e0e174fdc411b645548d92257ea

Change-Id: I2407e3bfb8a3709f9a9666d04c96efbd56469d92