Upgrade phan to 0.8.13
Final step to catch up with the 0.8 release. This notably requires a
higher version of php-ast (0.1.5), but adds support for PHP 7.2.x.
What's done here:
*Decl was replaced by Node (which is easy to fix)
*Adddressing https://github.com/phan/phan/commit/ea82b9948aa1263c5858ffbe780034bcd842e811
**Variables don't have a context anymore (see linked commit message),
and I couldn't find a way to link a variable to itself in the outer
scope. Thus I have removed such code - hoping that it won't break
anything.
**Variables now extend UnaddressableTypedElement instead of TypeElement,
so I have added this new class to all involved docblocks, and removed
the typehints as there's no common class/interface between the two.

Bug: T216974
Depends-On: I8645dc5636425ec0fd76d28bf6e51bfe816b0ab5
Change-Id: I33fac1d4a42464196c6b61d02ed8ed41cf413a99