Change inheritance for MW analyzer
Subclass the generic Visitor and PreVisitor instead of the Base one, and
adjust method calls to include a call to parent::method. This avoids a
double call to genericVisitor and MWVisitor, and thus simplifies the
switch to PluginV2.
Also remove memory usage print, as phan can already do it with the
print_memory_usage_summary option.

Depends-On: Ida41701efb0fb4e51be889749cc14459f14ff7df
Change-Id: I60882d39c5dc9b505d8e7eef7c67bc3d3aa47b4a