Fix phan issues
This patch fixes all issues reported by phan on the plugin itself. Some
of them are just nit-picky issues (mostly suppressed), while a couple of
them were actual bugs.
It also updates the config used for tests, adding a couple of plugins,
and removes old suppressions used for phan bugs which were resolved
since 0.8.0.
On top of that, avoid specifying php7.0 in integration tests, as we now
accept any version of PHP >= 7.0.0.

Depends-On: I3f29120ae1674cb542942666c89601e3d2df73a7
Change-Id: I44f43e67e652c2b88e38bc04d251b4a6757a85a0