Unbreak docblock parsing
The parsing of docblocks was broken during the update to 0.8.13. This
happened because their specific docComment class property has been
removed. However, in the codebase there's plenty of undeclared
properties, and the check for them is disabled in phan (and I have disabled it in PHPStorm, too). This makes it pretty hard to find buggy uses of undeclared class properties.

Depends-On: I44f43e67e652c2b88e38bc04d251b4a6757a85a0
Change-Id: Id6a6dbc1d8c6a743e69990aad328f776630a205c