Make nodeIsString work again
I left a long comment explaining the change. As I understand it, the
hasArrayLike check used to work fine just because of a phan bug (fixed
in [0]). Now that phan performs the proper check, the function broke.
I'm unsure whether not keeping a check for array-like is correct, but
this code is equivalent to the one using phan 0.8.0.

[0]: https://github.com/phan/phan/commit/a681f010d41f441ec7035063df219f8c5ae3a21b

Depends-On: I75a75d1e6487aa079889ba374803735f30ed2670
Change-Id: Ibf7e49caa812ee2f9b95f4af5a556d78ba73be88