Restore code bit for linking var to parentvar
This was removed in order to upgrade phan to 0.8.13 (see [0]), because
since then, Variables hold a FileRef and not a Context for better
performance, and thus it's not possible to get the scope of a var (see
[1], specifically the commit msg). While this is still true, that caused
a lot of false positive, so restore that code bit and retrieve the
external scope by calling getParentScope() on the context scope.
This is sometimes wrong, so an awful hack is added to make this bit of
code sort of work until we can upgrade to phan 2.0.0. A detailed
explanation is included in the inline comment, and once again, we really
need to get rid of that hack ASAP.

This patch also removes some outdated comments and adds a couple of
tests, given its hackyness level.

[0]: I33fac1d4a42464196c6b61d02ed8ed41cf413a99
[1]: https://github.com/phan/phan/commit/ea82b9948aa1263c5858ffbe780034bcd842e811

Depends-On: Ibf7e49caa812ee2f9b95f4af5a556d78ba73be88
Change-Id: I781e22f4b23fe1eacea628a42e90bdc8419d1ac1