Transfer the taintedness from objects to props
When trying to determine the taintedness of a property, check first if
the variable which hosts the property is tainted. If so, transfer its
taintedness to the property, keeping info about the original error
location. This is somehow similar to what's done for arrays.

Why is this change here:
Between phan 1.1.8 and 1.1.9, retrieving objects for a given node was
improved. Take the arraynumkey test, line 88. When we call visitProp for
the object, previously it found 0 props and we assigned a YES_TAINT
taintedness to it just because the variable is named $row. This is
crazily hacky, but it worked. After 1.1.9, phan returns a Property object for that bit of code, and we didn't set the appropriate taintedness to it.

Depends-On: Ia1002b898af4b95db2b92f23a4e6094f845c924b
Change-Id: I79619a5f47d5f2ba3ff0dc4ed515323935d43be9