Take into account implicit BranchScopes
See the comment in the code for a longer explanation. Basically, phan
has started replacing the FunctionLikeScope with a BranchScope for code
following a conditional return (only with a negated condition!). We need
to ensure that this works fine.
Moreover, ensure that taintedness on the same variable is overwritten in
case of reassignment if we're inside a branch scope (either explicit or
implicit).

Change-Id: Ia4ebb53ba3c5df038e81fabd282745a2a8c3190d