Handle nested calls
https://github.com/phan/phan/commit/42fea89e7b920a890e1b6d43765d7c48e730fc06
changed the way methods are analyzed, by avoiding analysis if phan deems
it useless and making the maximum recursion limit a static property. This reduced the overall maximum recursion depth, and we may get false positives if analysis is prevented.
Then,
https://github.com/phan/phan/commit/1dea1b4637b09b94b34c56a9ab6f36535057a7c5
added further optimization to avoid analysing methods in certain
situations, and this again could cause false positives.
I couldn't find an optimal solution to implement, so I just copied Analyzable::analyze and removed any performance check, to ensure that we always perform deep analysis when we need it.

Depends-On: Ia4ebb53ba3c5df038e81fabd282745a2a8c3190d
Change-Id: If90939dc4dfb99265992764539d3f03974396cb6