Add some unhandled node kinds
The ones added here are easy/not controversial. The one for AST_CLONE
should be expanded as noted in the todo, but that's something we can
postpone (like we do e.g. for new $variableExpr).
I ran seccheck on a couple of extensions to see if we're missing
something important, and it seems the important node kinds we're missing
are AST_UNPACK, AST_STATIC, AST_REF and AST_ASSIGN_REF.

Depends-On: Id06fdf04fbbee872559732bb9e8dee331fe7d38a
Change-Id: I5378456b77c967c5104e4078229befca4baf80c0