Prevent class props from sending taintedness too far
I explained in the code what kind of false positives would appear if we
don't make this change. For instance, MW core has a massive rate of
false positives due to this, mostly XSS coming from widely used
functions like User::newFromName. This change is not perfect, but should
be enough for the time being, and until we'll be able to handle props
more precisely.
Also remove comments about deleting method links, because they can and
will be reused in case of 'Multi' taint (at least).

Depends-On: I781e22f4b23fe1eacea628a42e90bdc8419d1ac1
Change-Id: Ia1002b898af4b95db2b92f23a4e6094f845c924b