Fix some issues with CI
This covers whitespace for phpcs, and restores integration tests via
shell instead of phpunit, as the latter is broken in CI due to absolute
paths. I'll play a bit with it after this patch is merged.
Also replace mysqli with a Database class in a test, as it's not
guaranteed that PHP will have the mysqli extension installed.

Change-Id: I00bf76099d3552ccf3c1b1541b9aacccd9f9d2ad