Upgrade phan to 2.0.0, ast to 1.0.1 and require PHP72+
This is the first step towards a seccheck 3.x release. The main
difference introduced with phan 2.0.0 is the upgrade to AST version 70;
which, in turn, notably introduced AST_CLASS_NAME for MyClass::class,
replacing AST_CLASS_CONST.

This patch also fixes a bug with $argv introduced in this version of
phan. Given that $argv can now be in the current scope as local
variable, the workaround introduced in
Ibfcac04a50f6daf68c4eeb2a3805d568665551c8 isn't working anymore. Hence,
rollback that change (i.e. normal globals aren't considered tainted
anymore, not by that part of code), and add special handling for $argv
and $argc. Doing that, we can consider them tainted regardless of
whether they're in the current scope or not. This doesn't hold true for
normal globals, because the local variable may just be a homonym (if the
global wasn't imported with the 'global' keyword).

This version could in theory support PHP 7.1, but let's break compatibility
now, since PHP7.1 is already EOLed.

Change-Id: I0af495eacc77fd71eb55b8a4025c8fd9a0e9a4c6