Remove explicit dependency on ext-ast
Phan doesn't depend on it anymore, nor do we. ext-ast is also suggested
in phan's composer.json, so we can avoid requiring it and make seccheck
easier to be installed.
If you try to run seccheck without php-ast, phan will exit and inform
you, and also suggest to allow the polyfill parser. However, this isn't really an option for seccheck (unless you have >4GB of RAM just for a MW extension).

Bug: T235053
Change-Id: If64a7c7a4481f9863137b0ed4c18271d30c2bed2