Add RAW_PARAM taint type
This is e.g. for Message::rawParams, and represents a type of taint that
isn't meant to be removed later on (i.e. the value should already be escaped). In practice, this is treated the same as EXEC (i.e. it fires a warning immediately when called, and not e.g. when output), except that it's not propagated.

Change-Id: I420d08b482346c91b6359679b96d717998db396f