Better handling of reassignments of the same var
Taint-check has never been able to properly understand reassignments of
the same var; this usually indicated poor code (variable re-use isn't
nice), and have been worked around in various repos. However, phan
2.2.13 makes this issue even more noticeable, probably because something
has increased the analysis depth. Hence, we need to improve this part
before moving on.

In order to achieve that, we stop updating the taintedness of local
variables when we find a call to an exec method: updating the
taintedness of the local variable associated to the function param will
inevitably create false positives ("assigning a tainted value..."), and
cause duplicate lines in caused-by. We don't need to do that because any
info about the taintedness of a local variable can and will be
determined when analyzing the body of the method, so no need to create
an incomplete duplicate. In theory, we should keep updating the taint of class
props and globals. However, in practice, props and globals can never end
up in that code path, hence I've left @fixme comments to highlight that.

Moreover, we stop linking the taintedness to the parent var in case the
latter turns out to be a parameter (not pass-by-ref). This caused
innocuous calls to echo and htmlspecialchars to be deemed unsafe when
analyzing the body of the function just because we had already analyzed
it with a tainted arg (e.g. 'samevar' test line 24 emitted a
"stand-alone" doubleescaped due to the call at line 14).

Change-Id: Id895591bd71bfde1b1603de099569003d73a6c3b