Further improvements for same var reassignments
This is another edge case, and it's causing various false positives in
LogFormatters. When reassigning something to a parameter, we are already
resetting the taint and avoid taint linking in most cases, but there's
one case left.

Based on the added test in samevar:

When we find htmlspecialchars($html), we know that $html is EXECed;
hence, we call ContextNode->getVariable and then markAllDependentMethodsExec with the object we got. The problem is that, despite the variable is reassigned above, phan gives us a Parameter object, which has its own taint links that we don't need. Thus, taint-check ignores the reassignment.

So, whenever we find an assignment where we override the taint, clear
all the links on the object, as they're no more necessary and can lead
to this kind of false positives.

However, we also have to exclude cases where a rhs object and a lhs
object are the same; unfortunately, due to an unrelated bug, we don't
handle the rhs properly if it's an array. As a side note, computing
$rhsObjs outside of the foreach makes the code a little bit faster.

As a side note: it took me a while to figure out why the test with
getEscaped3 was failing. Then I realized that taint-check is right, and
no double escaping was taking place!

Change-Id: I04209ee101d335f1d7abb847b780799e9df0ef03