Make handleMethodCall always require a FunctionInterface and a function FQSEN
If we had no function, this means that the call came from TaintednessVisitor
where we couldn't find a function. Well, based on the taint that we passed
in, handleMethodCall ended up skipping everything, and all it did in case of
no func was basically OR the taintedness of all args, add UNKNOWN_TAINT,
remove exec and return that. However, I don't think it's really correct to
return the taintedness of the args if we don't know what the method does.
Hence, just return UNKNOWN.

Change-Id: Icf119ff3e5257b8e0d68d26f7a931bc4aa49ca70