Avoid shelling out to run tests
On my machine, this is way faster than shelling out, and it's also saner.
Also avoid excluding the whole test folder from PHPCS and fix a couple
of issues there.

Note that we need to play a bit with backupStaticAttributes in order to
avoid any side effect due to caching.

Bonus: make phan-for-plugin always use the long progress bar, which is
shorter and doesn't clog CI logs.

Change-Id: I821e1eec97d93baff7550b7817d2e7b47f0e9c70