Fix another edge case interaction with phan
A partial explanation is at T249491#6032165. A shortened version is:
when we first analyze a method, we save the variables declared inside
and retrieve them back later. However, this didn't work for methods
called on subclasses, where the subclass didn't redefine the method
(see test case). In turn, this lead to a lack of variables for the
subclass call, which caused $this to be marked as undeclared.

Bug: T249491
Change-Id: I1c78ee6ce58de0da086a7bfc8ef4a7a2a25c4be4