Fix bad interaction with phan, part 3
This is closely related to T249491. This time, we try to retrieve the
"original" function scope before setting it, hence end up with no scope.
The fix implemented here is to just trigger an analysis of that method
in place. This should make sense, as otherwise, the analysis could be incomplete. This *might* cause some slowness, but that will be addressed as part of T203651.

Bug: T249519
Change-Id: I1fcf8182c48b6046bdf3950d795f8b7f2e1fcc98