Resolve 'static' type when searching hook handlers
This would previously fail in the attached example, where $this is used
as handler. Also avoid failing hard when a subscriber cannot be
found[1], and change a variable name because it may not strictly
represent a callback.

[1] - This copies the behaviour for other core methods introduced in
I29215598787db8fb2d13e3af693b24d33e97fab6. Perhaps we should emit a
non-security issue when this happens.

Bug: T250371
Change-Id: I87294afbf6b5c88fe014f05da592668d471aeef4