Upgrade to PluginV3
PluginV2 were removed in Phan 3 [1]. The only change is the addition of
PHP71 typehints, notably 'void' typehints. This requires changing the
visitors not to return anything.

Sometimes the new syntax is a bit harder to read (e.g. when we have
early returns), and using a class property can be also harder to
understand. On the bright side, some statements are shortened, and some
passthrough returns were removed. Either way, this is just a personal
preference, and we'll have to stick to void typehints, like it or not.

[1] - https://github.com/phan/phan/blob/master/NEWS.md#may-02-2020-phan-300-rc1

Change-Id: I2a42e4404d2b9e5c9b8bfe79b2f98f66817048c5