Track returned objects in-place, delete GetReturnObjVisitor
When we needed to collect the objects returned by a method, we would
previously visit that method using GetReturnObjVisitor, analyzing
the method if needed.

This patch proposes a different approach: whenever we find a return
statement during analysis, the object returned is saved in a dynamic
property inside the method. Retrieving the objects returned by a
method simply means reading that dynamic property. A reanalysis could
still be needed, but there's a guard for infinite recursion.

By design, in case of a method calling itself, this will return a
partial list of object. This is based on the assumption that partial is
better than nothing.

Practically speaking, this approach fixes a few edge cases with (I
think) subclassed methods, and it causes a dozen of issues to be
reported for core (and a couple to go away). I checked those issues, and
they're all in the maintenance/ directory. This means that they're false
positives, but taint-check doesn't know that (yet) as it assumes that
$argv is tainted. Overall, this translates into higher precision.
Curiously, these seem to be the same issues that disappeared with
Ib552d89f8dec84ff6c48176f7aa5bdbcc80423dc.

Performance-wise, there's no significant improvement. The time variation
is too small to be significative (GetReturnObjVisitor itself didn't have
a noticeable impact). The memory usage for MW core goes from
2136MB to 2125MB, probably because we now store single objects instead
of a Scope object containing them all.

Bug: T250671
Change-Id: Ie13ea3b83e10b70c0438924a2a22061b35404190