Track what taintedness is involved in all caused-by lines
This allows us to select only the interesting lines when looking up the
caused-by lines, whereas previously we would pick up any non-safe line.
This is especially evident in the attached test file, where the
caused-by line would include *all* methods call.

In the future we may use this info for other purposed, and/or do the
same for arg taint, too.

Note that the shape [ [ taint, line ], ... ] was chosen because it
allows to keep the lines ordered, while potentially allowing the same
line to appear with different taints. Currently we're not doing that for
perf reasons, but we might in the future.

Also, let me say that I'm personally very happy to see how better the
caused-by lines in tests are looking after this change!

Bug: T250679
Bug: T257351
Change-Id: I5a2bd0291d6cac69041b319f01ba8bed4639dc02