Big pass-by-ref overhaul
This patch revamps the handling of passbyrefs to fix a lot of common use cases.
The new algorithm is explained in TaintednessBaseVisitor::handlePassByRef, and
it's (IMHO) faster and easier to understand than the previous one. It's also
cheaper to maintain, and more robust because it shares many code paths with
non-ref variables/params/arguments.

On top of that, this patch fixes a very specific case with hooks: previously, we
wouldn't analyze the hook's subscribers, hence failed to collect various info
about taintedness. While previously this wasn't too big of an issue, it is now,
because we need phan to analyze methods with Parameters replaced with
PassByRefVariables in order to understand where the taintedness is going. In
fact, the 'json' and 'skinjson' tests were passing just because of an unrelated
bug in setTaintedness.

Finally, we now make sure that we'll never ever try to write taint data in
PassByReferenceVariable objects (they're too volatile for our needs), and we
extract the arg from them recursively.

This comes with a whole lot of tests, and old tests were renamed so it's
possible to filter with `strpos( $testName, 'passbyref' )`.

Bug: T250689
Change-Id: Ieb4c1610c352ad4ea11dc84640ca74458a4a64fc