Keep the line number up-to-date when visiting children nodes
This is especially relevant when emitting issues. For instance, take the
attached test. The issue at line 23 is emitted when visiting the call,
and that's fine. However, when we visit the array at line 22, and we
try to get its taint, we recurse into the children of the array, i.e. the
function call. So we analyze the call again, BUT the context is still at
line 22, so we emit an issue again at line 22.

The (obvious) solution is to update the context line number before visiting a
new node with the node's line number.

This patch makes 26 redundant issues go away in MW core.

Note: this patch also includes a phan-interaction test for an issue that
was spotted during early tests of the patch (and only happened when
cloning the context before changing the line number).

Change-Id: I92795d963039ec22e0252e1897c84d56fcc530c3