Simplify handling of (pre|post)(increment|decrement)
The 'var' children is guaranteed to exist (after all, you're supposed to
use these operators to change the value of something...), so there's no
need to check for its existence. However, it's not guaranteed to be a
variable, hence using visitVar was wrong. On top of that, there's no
need to see if we can extract a phan object and then return its
taintedness; using getTaintedness() can handle that case, as well as
literals, as well as cases where getPhanObjsForNode returns nothing.

Change-Id: Ic3c1ff96f9c9c6458a4bbaedddfe33f3ba4703a3