Partly unify binop handling with assignop handling
We now use the correct mask when dealing with assignops (e.g. .=),
copying what we do with normal binops. This is still excluding binary
addition, because that needs to be account for arrays etc.

Note: this patch causes some "good" caused-by lines to
go away, e.g. in the indirectecho test. Those lines weren't actually
meant to be there (it was an unintended, albeit positive, side effect),
and in fact they go away if you replace the '.=' with '='.

Change-Id: I76fef486ce82f3312438d3eae1d557502984ba00