Backpropagate SQL_NUMKEY taint
This must be special-cased for it to work. The implementation here
prefers false negatives over false positives. Specifically, it doesn't
backpropagate taint of function arguments (e.g. `array_merge( $unsafe,
$whatever )`), but that's the same for taints other than numkey.

The approach taken here prefers false negatives, and will not
backpropagate the taint if the argument is a variable (or a prop) with
mixed keys. Trying to do so would currently result in taintedness being
propagated to safe (aka string) keys as well.

This patch doesn't add any new error in MW core, which all in all is a
good sign (that there are no obvious SQLi's).

Change-Id: I5ad4e78e127d51e030db11cef8404f7c8fc5cc07