Hook into ContextMergeVisitor to merge taint data, bump phan/phan to 3.0.3
This will dramatically improve taint-check's ability to read conditional
branches. Basically, the situation is:
 - Previously, we kept a single copy of every variable, merging any new
 taint on the same object for conditional branches;
 - We now have different copies for different branches; each copy will
 only be used in its branch, but at the end of the ramification, we'll
 merge taint data for all branches in the "parent" object.

On MW core, roughly 30 false positives go away with this change.

This change has a huge impact, and it's one of the long overdue
refactorings that can improve taint-check's behaviour in lots of
situations. I'll let the tests vouch for this.

Note that this capability was introduced in phan 3.0.3 (see
https://github.com/phan/phan/issues/3956)

Change-Id: I7a22a3b5edd466674f14ce2bf3e1a74ee64fa218