Improve and simplify nodeIs* methods
 - Add a helper to get the type of a node
 - Normalize docblocks to 'Node|mixed' (here, this means either a Node
 or a scalar from the AST)
 - Avoid using hasType ( FooBarType::instance() ), in favour of
 dedicated accessors.

Specifically, using accessors means that type inference is now improved,
because e.g. if a UnionType only consists of LiteralIntType, then we
have

 $type->hasType( IntType::instance() ) // false
 $type->hasIntType() // true

and this covers the cases where $node is "almost" a literal (i.e. not a
literal, but something that we know 100% refers to a literal).
In turn, this improves a bit the handling of NUMKEY with non-literal
keys (previously taint-check would not recognize the second example as
unsafe).

Change-Id: I4c67c892dea95740e11cf5cdb4cd41725ea9938a