Store taintedness in a value object
For now, this has little benefit (it just helps a bit with strong typing and
code readability). However, this is a first step towards introducing
more complicated "types" of taintedness. Notably, I plan to use this for
introducing offset taintedness (T253875), which should coexist with the
"classic" taintedness.

Re performance: a test run on core showed that the runtime is likely not
affected (it went from 2m3s to 2m8s, but this isn't significant
IMHO), but the memory usage is (as expected). In fact, it goes from
2679MB to 2696MB, which isn't too bad anyway (considering that we create
a Taintedness object for every clone of every variable/property in the
codebase).

Change-Id: I1684a98c5dae66e9314212bce05b8de2015aaf8b