Introduce a value object for function taintedness
This is a follow-up to the introduction of the Taintedness class, and it
helps clean up a few places. The benefits, for now, are the same as the
Taintedness class, but they'll be more evident in the near future. This
commit also includes a minor improvement, i.e. caching of builtin func
taints.

A test run on core gave the following results:
 - Runtime 1m59s
 - Memory 2691MB
which means lower memory usage (runtime variation is not significant)
compared to the parent (I1684a98c5dae66e9314212bce05b8de2015aaf8b).

Change-Id: I3f603e12a65fb76b51e51f85fd6b8a5337c4b48e