Move CLI XSS section from "limitations" to "false positives"
This isn't a limitation per se, because we cannot expect the plugin to
tell that the file it's analyzing will only be executed in CLI context.

So insead of an unhelpful "limitation" message, move it to the "how to
avoid false positives" section and explain how it can be avoided.

Change-Id: I3ffcb124a9ee169a7b3bd1dececdfd4c0578cb8f