Readability: rewrite all taint constants using left shift
This might not be useful for 1, 2, etc., but it's certainly useful for
anything above 1000. Also, use combination of other constants instead of
magic values for catch-all taints.

While at it, fix a bug in a regexp that caused raw_param to be treated
as a normal taint type (the only consequence of this was that the plugin
would accept a prefix for it, which would've ended up not being
recognized anyway).

Also, swap NO_OVERRIDE and RAW_PARAM, as the former is 2^29 and the latter is 2^30.

Change-Id: I2ea0abf5d79d8f1789720e2551e0343c7977d541