Fix bug which made variables lose their type if overridden in branches
This was probably caused by the plugin triggering an anaylysis of the if
condition *after* the body was analyzed (in post-order). But after
analyzing the body, $ts has another type, and thus the analysis fails.

We can make this a noop because, as noted in the comment, everything
inside the if will be analyzed, plus an if doesn't have any taintedness
per se (it's not treated as an expression).

Bug: T268891
Change-Id: Ieff16212eb72970b50f0928575feece043fab16b