(MW) Improve detection of tainted HTMLForm labels
With key taintedness being tracked separately, we can now greatly
improve detection of tainted HTMLForm labels, because we can discard the
taintedness of the values.

The code can also be simplified, at the cost of removing indications
about the value corresponding to the tainted key in the literal case.
However, this is probably wasn't very useful, as values are often
non-literal.

Additionally, try resolving every field name/value, as that will make
the detection work even if they were specified indirectly in a way that
is easy to infer statically.

The limitation mentioned in the README also doesn't exist anymore, so
remove that line.

Change-Id: I44ba9dc6be6b3c989f35f491387afdb1ed854f6a