(MW) Slightly improve the method for handling Database::makeList
Always try to resolve the LIST_* param, clean up the code. The only
functional change is that it now detects cases when the constant is
first stored in a variable, or retrieved in another way that can be
determined statically 100% of the times, which is probably never going
to happen, as you might as well just pass a literal at that point.

Also add some logging for when the value was passed literally and is not
a valid one.

Change-Id: I0af1a2f8428032c78ad89078fa290d91346efcc0