Remove $override in the case of assign ops
This is essentially equivalent to the previous code. Instead of
overriding, we merge the LHS taint into the RHS, which we were already
doing. Then this overall taint is used everywhere. The only exception
should be when determining the error taint, as that should only have the
RHS's own taint, but doing so now would change some caused-by lines, so
it's left for later.

Change-Id: Ic9fe996ee45faeb957056b81e169df36943c991b