Store arg errors in arrays, like other caused-by lines
This completes the migration started with
Ia6190b341e5846269a2e0071e84431eeab40ee60, and all caused-by lines are
now stored in arrays, together with the taintedness of the specific
line.

Additionally, use mergeCausedByLines when retrieving the lines, not
array_merge, and filter out extra lines at the end.

Finally, fix the conditional for checking whether the line we're adding
was already present -- the previous code was checking whether an
old-style string error was in an array of array errors.

Change-Id: I462cf39b9665dbce6fdc9c7a14a79b4331202a38