Don't link parameters of functions with hardcoded taintedness
Previously, we would link parameters to the function so we could
propagate side effects and whatnot. This has two major disadvantages:
- It slows down analysis, because there are more dependencies to track.
  Note that these dependencies cause relatively frequent reanalysis,
  because hardcoded methods are usually widely-used methods (e.g.
  methods in Database, or in Html). The difference for MW core is very
  tiny (roughly 5 seconds and 20MB), but not invisible.
- It makes caused-by lines propagate when they shouldn't; for instance,
  in the attached test there were lots of lines related to
  Message::text, OutputPage::addHTML, unrelated calls to rawElement and
  some code inside the Html class. These lines made the output hard to
  understand, and in real-world usage can easily hide useful caused-by
  lines because the length limit is reached before finding a useful
  line.

OTOH, disabling linking for builtin functions has basically no
downsides, precisely because anything dangerous was already hardcoded.
The result on an old checkout of core is roughly the same, with a bunch
of issues now having shorter caused-by lines, a few false positives
coming and going, and a couple of true positive DoubleEscaped being
reported.

Bug: T247428
Change-Id: I786bc1de5b071c0d660b6cf5be1834e9a47c42bc