Better tracking of shapes in isSafeAssignment
The current code only works because most taintedness values at the LHS
don't have a shape. Assuming that this is going to change, we now
properly account for shapes when determining if an assignment is safe.

This patch should in fact be a noop for now, as the analysis of MW core
is unchanged in terms of issues and memory.

Change-Id: Iad08964bda6373e7baf9f73a6a23c01749e785af