Simplify hook registration
Phan already provides methods for retrieving a callable from a node, so
use them. This also avoids several bugs (such as not working when the
class is namespaced) and limitations (e.g. it can now understand some
non-literal callables, as long as the value can still be inferred
statically, e.g. if it's assigned to a variable first).

Change-Id: Ia7921b921ee8805c7c2f63c9f4e6ce1d24cd2b31