Implement caching for node taintedness
This is similar to phan's caching of union types. The cache hit ratio is
around 94% (approx 807000/856000 on MW core), as for performance:
- Taint-check would consume 320 MB of additional memory
- The runtime decreases by about 30-35 seconds

Another advantage of doing this is that in some situations, we correctly
read the old data, and not the future data. For instance, in
  $var = 2;
  echo $var = 'foo';

when analyzing the second assignment, the plugin would use "foo" as
value of $var, but since we're still analyzing the assignment, the value
is 2. I tried to come up with a test case to see if this really prevents
any bug, but I couldn't. Anyway it's more correct now.

Not only that: it avoids having to analyze children of a given node in
an order that phan doesn't expect, which can cause issues like
https://github.com/phan/phan/issues/4368 (see
I3a73e27a6abb5084009917c2c070bc0f115f0e5b).

Practically speaking, this commit makes 5 new errors appear for MW core.
AFAICS these are indeed false positives, but in such a way that the
plugin cannot really tell (e.g. one is sanitizing data with %d in
sprintf, 3 of them are using the fact that a specific DB column only contains
safe values, and for the last one, taint-check is getting confused with
array keys). All in all, this is a good sign.

For now, caching is only implemented in cases where computing the
taintedness is now immediate (i.e. INAPPLICABLE nodes were skipped).

Change-Id: I1301c27425ed9906a38b712cfcd7b3709433c2c9