Create a separate visitor for backpropagating taintedness
For now, this just copies code from getPhanObjsForNode, with 'return' as
the only option. Later on, it will be used to change the taintedness
before recursing, for more precise results. Also, more generally, it
might be expanded with paths that only make sense in the context of
backpropagating taintedness (since getPhanObjsForNode is quite generic).

The numkey case still uses the old code, because eventually numkey
should just become a special-case in this visitor, handled on the fly.

Change-Id: I0e9a354d0ec43e29263c75323a82eb8d99defc14